@@ -1,11 +1,18 @@
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
+import { NgModule } from "@angular/core";
+import { Routes, RouterModule } from "@angular/router";
+import { LoginComponent } from "./login/login.component";
+import { RegisterComponent } from "./register/register.component";
+import { UserComponent } from "./user/user.component";
 
-
-const routes: Routes = [];
+const routes: Routes = [
+  { path: "login", component: LoginComponent },
+  { path: "register", component: RegisterComponent },
+  { path: "user/management", component: UserComponent },
+  { path: "", redirectTo: "/login", pathMatch: "full" },
+];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],
-  exports: [RouterModule]
+  exports: [RouterModule],
 })
-export class AppRoutingModule { }
+export class AppRoutingModule {}
