@@ -68,7 +68,8 @@ export class AuthenticationService {
     this.loadToken();
     if (this.token != null && this.token !== "") {
       if (
-        (this.jwtHelper.decodeToken(this.token).sub != null || "") &&
+        this.jwtHelper.decodeToken(this.token).sub != null ||
+        "" ||
         !this.jwtHelper.isTokenExpired(this.token)
       ) {
         this.loggedInUsername = this.jwtHelper.decodeToken(this.token).sub;
